nightly_eval/requirements.txt
boto3==1.34.0
pandas==2.1.4
pyarrow==15.0.0
psycopg2-binary==2.9.9
mlflow==2.11.0
numpy==1.26.4
